refactor(test): refactor test
